Consolidate SymDispatchMode into ProxyTensorMode
Instead of having a separate context variable for SymDispatchMode, we
now simply delegate to the current active proxy tensor mode when we
need to trace a SymInt.  We maintain a separate `__sym_dispatch__` magic
method as the calling convention is different than `__torch_dispatch__`.

Consolidating the modes in this ways means that we can consistently
disable both of these modes in tandem simply by removing the mode
from the proxy mode infra slot.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>

ghstack-source-id: 59d1d82a4300e1926f19b70358f6637cc01319dd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/132674